[refactor] #40 split protocol tests
