+++ PlanetaMatchMakerServerTest/PlanetaMatchMakerServerTest.vcxproj.filters
@@ -2,7 +2,13 @@
 <Project ToolsVersion="4.0" xmlns="http://schemas.microsoft.com/developer/msbuild/2003">
   <ItemGroup>
     <ClCompile Include="main.cpp" />
-    <ClCompile Include="protocol_tests\message_handlers_protocol_test.cpp" />
+    <ClCompile Include="protocol_tests\authentication_protocol_test.cpp" />
+    <ClCompile Include="protocol_tests\connection_test_protocol_test.cpp" />
+    <ClCompile Include="protocol_tests\create_room_protocol_test.cpp" />
+    <ClCompile Include="protocol_tests\join_room_protocol_test.cpp" />
+    <ClCompile Include="protocol_tests\list_room_protocol_test.cpp" />
+    <ClCompile Include="protocol_tests\message_flow_protocol_test.cpp" />
+    <ClCompile Include="protocol_tests\update_room_status_protocol_test.cpp" />
     <ClCompile Include="unit_tests\checked_static_cast_test.cpp" />
     <ClCompile Include="unit_tests\datetime_test.cpp" />
     <ClCompile Include="unit_tests\errors_test.cpp" />
@@ -19,5 +25,8 @@
     <ClCompile Include="unit_tests\session_data_test.cpp" />
     <ClCompile Include="unit_tests\thread_safe_data_container_test.cpp" />
     <ClCompile Include="unit_tests\utilities_test.cpp" />
+  </ItemGroup>
+  <ItemGroup>
+    <ClInclude Include="protocol_tests\protocol_test_support.hpp" />
   </ItemGroup>
 </Project>
