+++ PlanetaMatchMakerServerTest/protocol_tests/authentication_protocol_test.cpp
@@ -1,0 +1,139 @@
+#include <boost/test/unit_test.hpp>
+
+#include "protocol_test_support.hpp"
+
+namespace {
+	using namespace pgl::test;
+}
+
+BOOST_AUTO_TEST_SUITE(authentication_protocol_test)
+	BOOST_AUTO_TEST_CASE(test_authentication_request_replies_success_and_assigns_player) {
+		protocol_context context;
+		const pgl::authentication_request_message request{
+			pgl::api_version,
+			pgl::game_id_t(context.setting.authentication.game_id),
+			pgl::game_version_t(context.setting.authentication.game_version),
+			u8"player"
+		};
+		protocol_handler_run handler(context, pgl::message_type::authentication);
+
+		write_packed(context.client_socket, pgl::request_message_header{pgl::message_type::authentication}, request);
+		const auto reply_header = read_packed<pgl::reply_message_header>(context.client_socket);
+		const auto reply = read_packed<pgl::authentication_reply_message>(context.client_socket);
+		const auto exception = handler.wait();
+
+		BOOST_CHECK(!exception);
+		BOOST_CHECK(reply_header.message_type == pgl::message_type::authentication);
+		BOOST_CHECK(reply_header.error_code == pgl::message_error_code::ok);
+		BOOST_CHECK(reply.result == pgl::authentication_result::success);
+		BOOST_CHECK_EQUAL(reply.player_tag, 1);
+		BOOST_CHECK(context.session_data.is_authenticated());
+		BOOST_CHECK(context.session_data.client_player_name().name == u8"player");
+		BOOST_CHECK(context.server_data.get_player_name_container().is_player_exist(
+			context.session_data.client_player_name()));
+	}
+
+	BOOST_AUTO_TEST_CASE(test_authentication_request_replies_game_id_mismatch_and_disconnects) {
+		protocol_context context;
+		const pgl::authentication_request_message request{
+			pgl::api_version,
+			u8"wrong-game",
+			pgl::game_version_t(context.setting.authentication.game_version),
+			u8"player"
+		};
+		protocol_handler_run handler(context, pgl::message_type::authentication);
+
+		write_packed(context.client_socket, pgl::request_message_header{pgl::message_type::authentication}, request);
+		const auto reply_header = read_packed<pgl::reply_message_header>(context.client_socket);
+		const auto reply = read_packed<pgl::authentication_reply_message>(context.client_socket);
+		const auto exception = handler.wait();
+
+		BOOST_CHECK(is_intended_disconnect(exception));
+		BOOST_CHECK(reply_header.message_type == pgl::message_type::authentication);
+		BOOST_CHECK(reply_header.error_code == pgl::message_error_code::ok);
+		BOOST_CHECK(reply.result == pgl::authentication_result::game_id_mismatch);
+		BOOST_CHECK(!context.session_data.is_authenticated());
+	}
+
+	BOOST_AUTO_TEST_CASE(test_authentication_request_replies_api_version_mismatch_and_disconnects) {
+		protocol_context context;
+		const pgl::authentication_request_message request{
+			static_cast<pgl::api_version_type>(pgl::api_version + 1),
+			pgl::game_id_t(context.setting.authentication.game_id),
+			pgl::game_version_t(context.setting.authentication.game_version),
+			u8"player"
+		};
+		protocol_handler_run handler(context, pgl::message_type::authentication);
+
+		write_packed(context.client_socket, pgl::request_message_header{pgl::message_type::authentication}, request);
+		const auto reply_header = read_packed<pgl::reply_message_header>(context.client_socket);
+		const auto reply = read_packed<pgl::authentication_reply_message>(context.client_socket);
+		const auto exception = handler.wait();
+
+		BOOST_CHECK(is_intended_disconnect(exception));
+		BOOST_CHECK(reply_header.error_code == pgl::message_error_code::ok);
+		BOOST_CHECK(reply.result == pgl::authentication_result::api_version_mismatch);
+		BOOST_CHECK_EQUAL(reply.api_version, pgl::api_version);
+	}
+
+	BOOST_AUTO_TEST_CASE(test_authentication_request_replies_game_version_mismatch_and_disconnects) {
+		protocol_context context;
+		const pgl::authentication_request_message request{
+			pgl::api_version,
+			pgl::game_id_t(context.setting.authentication.game_id),
+			u8"2.0.0",
+			u8"player"
+		};
+		protocol_handler_run handler(context, pgl::message_type::authentication);
+
+		write_packed(context.client_socket, pgl::request_message_header{pgl::message_type::authentication}, request);
+		const auto reply_header = read_packed<pgl::reply_message_header>(context.client_socket);
+		const auto reply = read_packed<pgl::authentication_reply_message>(context.client_socket);
+		const auto exception = handler.wait();
+
+		BOOST_CHECK(is_intended_disconnect(exception));
+		BOOST_CHECK(reply_header.error_code == pgl::message_error_code::ok);
+		BOOST_CHECK(reply.result == pgl::authentication_result::game_version_mismatch);
+		BOOST_CHECK(reply.game_version == pgl::game_version_t(context.setting.authentication.game_version));
+	}
+
+	BOOST_AUTO_TEST_CASE(test_authentication_request_replies_parameter_error_for_empty_player_name) {
+		protocol_context context;
+		const pgl::authentication_request_message request{
+			pgl::api_version,
+			pgl::game_id_t(context.setting.authentication.game_id),
+			pgl::game_version_t(context.setting.authentication.game_version),
+			u8""
+		};
+		protocol_handler_run handler(context, pgl::message_type::authentication);
+
+		write_packed(context.client_socket, pgl::request_message_header{pgl::message_type::authentication}, request);
+		const auto reply_header = read_packed<pgl::reply_message_header>(context.client_socket);
+		const auto exception = handler.wait();
+
+		BOOST_CHECK(is_intended_disconnect(exception));
+		BOOST_CHECK(reply_header.error_code == pgl::message_error_code::request_parameter_wrong);
+		expect_no_more_reply_data(context.client_socket);
+	}
+
+	BOOST_AUTO_TEST_CASE(test_authentication_request_replies_operation_invalid_when_already_authenticated) {
+		protocol_context context;
+		context.session_data.set_authenticated();
+		const pgl::authentication_request_message request{
+			pgl::api_version,
+			pgl::game_id_t(context.setting.authentication.game_id),
+			pgl::game_version_t(context.setting.authentication.game_version),
+			u8"player"
+		};
+		protocol_handler_run handler(context, pgl::message_type::authentication);
+
+		write_packed(context.client_socket, pgl::request_message_header{pgl::message_type::authentication}, request);
+		const auto reply_header = read_packed<pgl::reply_message_header>(context.client_socket);
+		const auto exception = handler.wait();
+
+		BOOST_CHECK(is_intended_disconnect(exception));
+		BOOST_CHECK(reply_header.error_code == pgl::message_error_code::operation_invalid);
+		expect_no_more_reply_data(context.client_socket);
+	}
+
+BOOST_AUTO_TEST_SUITE_END()
